backend/pyproject.toml
# pyproject.toml

[project]
name = "text-archive-assistant-backend"
version = "0.1.0"

dependencies = [
    "fastapi==0.103.1",
    "uvicorn[standard]==0.23.2",
    "pydantic-settings==2.0.3",
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.21",
    "redis==4.6.0",
    "python-multipart==0.0.6",
    "alembic==1.12.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "paddleocr==2.7.0",
    "paddlepaddle==2.5.2",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "hypothesis==6.92.1",
    "httpx==0.25.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
namespaces = false
